+++ DutchiePay/src/main/java/dutchiepay/backend/global/payment/exception/PaymentErrorCode.java
@@ -21,9 +21,10 @@ public enum PaymentErrorCode {
     /**
      * 500 INTERNAL_SERVER_ERROR
      */
+    PORTONE_CANCEL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "결제 취소에 실패하였습니다."),
+    INVALID_KAKAO_APPROVE_RESPONSE(HttpStatus.INTERNAL_SERVER_ERROR, "카카오페이 결제 승인 응답이 유효하지 않습니다."),
     ERROR_KAKAOPAY_STATUS(HttpStatus.INTERNAL_SERVER_ERROR, "카카오페이 결제 상태를 확인할 수 없습니다."),
-    ERROR_KAKAOPAY_CANCEL(HttpStatus.INTERNAL_SERVER_ERROR, "카카오페이 결제 취소 중 오류가 발생하였습니다."),
-    PORTONE_CANCEL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "결제 취소에 실패하였습니다.");
+    ERROR_KAKAOPAY_CANCEL(HttpStatus.INTERNAL_SERVER_ERROR, "카카오페이 결제 취소 중 오류가 발생하였습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
